+++ lib.php
@@ -127,6 +127,7 @@ function theme_adaptable_process_scss($scss, $theme) {
     $defaults = array(
         '[[setting:linkcolor]]' => '#51666C',
         '[[setting:linkhover]]' => '#009688',
+        '[[setting:dimmedtextcolor]]' => '#6a737b',
         '[[setting:maincolor]]' => '#3A454b',
         '[[setting:backcolor]]' => '#FFFFFF',
         '[[setting:regionmaincolor]]' => '#FFFFFF',
@@ -500,6 +501,56 @@ function theme_adaptable_set_customcss($css, $customcss) {
 }
 
 /**
+ * Serves the H5P Custom CSS.
+ *
+ * @param string $filename The filename.
+ * @param theme_config $theme The theme config object.
+ */
+function theme_adaptable_serve_hvp_css($filename, $theme) {
+    global $CFG, $PAGE;
+    require_once($CFG->dirroot.'/lib/configonlylib.php'); // For min_enable_zlib_compression().
+
+    $PAGE->set_context(context_system::instance());
+    $themename = $theme->name;
+
+    $content = theme_adaptable_get_setting('hvpcustomcss');
+    $md5content = md5($content);
+    $md5stored = get_config('theme_'.$themename, 'hvpccssmd5');
+    if ((empty($md5stored)) || ($md5stored != $md5content)) {
+        // Content changed, so the last modified time needs to change.
+        set_config('hvpccssmd5', $md5content, 'theme_'.$themename);
+        $lastmodified = time();
+        set_config('hvpccsslm', $lastmodified, 'theme_'.$themename);
+    } else {
+        $lastmodified = get_config('theme_'.$themename, 'hvpccsslm');
+        if (empty($lastmodified)) {
+            $lastmodified = time();
+        }
+    }
+
+    // Sixty days only - the revision may get incremented quite often.
+    $lifetime = 60 * 60 * 24 * 60;
+
+    header('HTTP/1.1 200 OK');
+
+    header('Etag: "'.$md5content.'"');
+    header('Content-Disposition: inline; filename="'.$filename.'"');
+    header('Last-Modified: '.gmdate('D, d M Y H:i:s', $lastmodified).' GMT');
+    header('Expires: '.gmdate('D, d M Y H:i:s', time() + $lifetime).' GMT');
+    header('Pragma: ');
+    header('Cache-Control: public, max-age='.$lifetime);
+    header('Accept-Ranges: none');
+    header('Content-Type: text/css; charset=utf-8');
+    if (!min_enable_zlib_compression()) {
+        header('Content-Length: '.strlen($content));
+    }
+
+    echo $content;
+
+    die;
+}
+
+/**
  * Set display of course contacts on front page tiles
  * @param string $css
  * @param string $display
@@ -633,6 +684,8 @@ function theme_adaptable_pluginfile($course, $cm, $context, $filearea, $args, $f
             return $theme->setting_file_serve('frontpagerendererdefaultimage', $args, $forcedownload, $options);
         } else if ($filearea === 'headerbgimage') {
             return $theme->setting_file_serve('headerbgimage', $args, $forcedownload, $options);
+        } else if ($filearea === 'hvp') {
+            theme_adaptable_serve_hvp_css($args[1], $theme);
         } else if ($filearea === 'loginbgimage') {
             return $theme->setting_file_serve('loginbgimage', $args, $forcedownload, $options);
         } else if (preg_match("/^p[1-9][0-9]?$/", $filearea)) {